test: drain SPAWN_ALLOWED — no real process spawns left in unit scope
Migrate all 13 grandfathered Rule-5 spawner files off real process
spawns. Per-file strategy:

- In-process harness (runCliCapture): wiki.test.ts,
  distill-cli-flag.test.ts, plus the unit remainders of the split files
- Split (unit remainder + spawn tests moved to tests/integration/):
  remember-frontmatter → integration/remember-stdin
  env-path-run → integration/env-run
  secret → integration/secret-stdin
  secret-path-run → integration/secret-run
  commands/events → integration/events-offset-crossproc
  commands/improve-result-to-file → integration/improve-cli-result-storage
  commands/show-argv → integration/show-argv-entrypoint
  commands/improve-cli-flags → integration/improve-cli-flags
  env → integration/env-export-bash
  index-writer-lock → integration/index-writer-lock-crossproc
- Whole-file move: save-command → tests/integration/save-command
  (17 real-git call sites; integration by nature)

lint-tests-isolation: SPAWN_ALLOWED is now empty (set + rule kept),
ratchet baseline lowered 77 → 64 — the 2026-07-02 Rule-5 step-up was
drained same-day. biome.json noTemplateCurlyInString override extended
to the moved integration/env-run.test.ts (intentional shell ${...}
placeholders).

Gate: bun run lint clean, tsc --noEmit clean, unit 4972/0 across 8
shards, integration 732 pass / 0 fail.
