+++ biome.json
@@ -22,6 +22,7 @@
       "includes": [
         "tests/env-secret-tokens.test.ts",
         "tests/env-path-run.test.ts",
+        "tests/integration/env-run.test.ts",
         "tests/commands/env-set-unset.test.ts"
       ],
       "linter": { "rules": { "suspicious": { "noTemplateCurlyInString": "off" } } }
