+++ tests/commands/improve-result-to-file.test.ts
@@ -14,8 +14,7 @@
  */
 
 import { Database } from "bun:sqlite";
-import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { spawnSync } from "node:child_process";
+import { afterEach, describe, expect, test } from "bun:test";
 import fs from "node:fs";
 import path from "node:path";
 import type { AkmImproveResult } from "../../src/commands/improve/improve";
@@ -29,25 +28,12 @@ import { type SandboxedDir, makeStashDir as sandboxMakeStashDir, sandboxXdgDataH
 // The pure-function tests (buildImproveRunId, relativeImproveResultPath,
 // writeImproveResultFile) run in-process — writeImproveResultFile isolates
 // state.db via the allowlisted sandboxXdgDataHome helper. The three `akm
-// improve` CLI tests, however, run `improve` for real, which opens and WRITES
-// the state.db improve_runs table.
-//
-// HARNESS GAP — KEPT AS A SUBPROCESS: in-process, a SQLite write lock on
-// state.db is already held within the test process (the suite keeps DB handles
-// open across the run), so an in-process improve aborts with "state DB
-// busy/locked after retries" — a genuine process-level contention a fresh
-// subprocess does not have. Until the harness grows a way to release/clear all
-// state.db handles before an improve-executing call, those three tests spawn a
-// real `bun src/cli.ts` so each gets a clean, uncontended DB. The runCli helper
-// passes env to spawnSync rather than mutating process.env, so it does not
-// affect the in-process tests; it still mints a fresh XDG_DATA_HOME per call
-// (via the allowlisted sandbox helper) and returns that dir so the assertions
-// can read the improve_runs rows back.
+// improve` CLI tests that used to live here run `improve` for real (which
+// opens and WRITES the state.db improve_runs table, hitting genuine
+// cross-process SQLite contention in-process) and were moved to
+// tests/integration/improve-cli-result-storage.test.ts.
 
 const disposers: Array<{ cleanup: () => void }> = [];
-
-const repoRoot = path.resolve(import.meta.dir, "..", "..");
-const cliPath = path.join(repoRoot, "src", "cli.ts");
 
 function makeStashDir(): string {
   const stash: SandboxedDir = sandboxMakeStashDir();
@@ -57,36 +43,6 @@ function makeStashDir(): string {
   }
   disposers.push(stash);
   return stash.dir;
-}
-
-interface CliRun {
-  status: number;
-  stdout: string;
-  stderr: string;
-  xdgData: string;
-}
-
-function runCli(args: string[], stashDir: string): CliRun {
-  // Fresh XDG_DATA_HOME per call so each run writes its own state.db. Use the
-  // allowlisted sandbox helper to keep mkdtempSync out of the test file; the
-  // dir is passed to spawnSync's env (not process.env) so it never leaks.
-  const data = sandboxMakeStashDir();
-  disposers.push(data);
-  const result = spawnSync("bun", [cliPath, ...args], {
-    encoding: "utf8",
-    timeout: 60_000,
-    env: {
-      ...process.env,
-      AKM_STASH_DIR: stashDir,
-      XDG_DATA_HOME: data.dir,
-    },
-  });
-  return {
-    status: result.status ?? 1,
-    stdout: result.stdout ?? "",
-    stderr: result.stderr ?? "",
-    xdgData: data.dir,
-  };
 }
 
 /**
@@ -267,91 +223,5 @@ describe("writeImproveResultFile", () => {
     } finally {
       dataSb.cleanup();
     }
-  });
-});
-
-describe("akm improve CLI: result-to-state.db default + --json-to-stdout escape hatch", () => {
-  let stashDir: string;
-  beforeEach(() => {
-    stashDir = makeStashDir();
-  });
-
-  test("default mode records the full result in state.db improve_runs and emits NOTHING on stdout", () => {
-    // Use --dry-run so the run completes quickly without LLM calls.
-    const result = runCli(["improve", "--dry-run"], stashDir);
-    expect(result.status).toBe(0);
-
-    // Stdout is empty — no JSON summary, no envelope, no "result written to" hint.
-    expect(result.stdout).toBe("");
-
-    // The result row landed in state.db with the full body in result_json.
-    const rows = readImproveRuns(result.xdgData);
-    expect(rows.length).toBe(1);
-    expect(rows[0].ok).toBe(1);
-    expect(rows[0].dry_run).toBe(1);
-    expect(rows[0].result.ok).toBe(true);
-    expect(rows[0].result.dryRun).toBe(true);
-    expect(rows[0].result.memorySummary).toBeDefined();
-    expect(rows[0].result.plannedRefs).toBeDefined();
-
-    // No legacy on-disk artifact file is authored anymore.
-    const runsDir = path.join(stashDir, ".akm", "runs");
-    expect(fs.existsSync(runsDir)).toBe(false);
-
-    // No "improve result written to" hint on stderr — the existing [improve]
-    // log lines from improve.ts are the canonical console UX.
-    expect(result.stderr).not.toContain("improve result written to");
-  });
-
-  test("--json-to-stdout restores prior behaviour: full JSON on stdout, no state.db row written", () => {
-    const result = runCli(["improve", "--dry-run", "--json-to-stdout"], stashDir);
-    expect(result.status).toBe(0);
-
-    // Stdout has the full result body.
-    const parsed = JSON.parse(result.stdout) as Record<string, unknown>;
-    expect(parsed.ok).toBe(true);
-    expect(parsed.dryRun).toBe(true);
-    expect(parsed.memorySummary).toBeDefined();
-    expect(parsed.plannedRefs).toBeDefined();
-    // No envelope-only fields in legacy mode.
-    expect(parsed.runId).toBeUndefined();
-    expect(parsed.resultPath).toBeUndefined();
-    expect(parsed.summary).toBeUndefined();
-
-    // No improve_runs row written in --json-to-stdout mode.
-    const rows = readImproveRuns(result.xdgData);
-    expect(rows.length).toBe(0);
-
-    // No legacy on-disk file either.
-    const runsDir = path.join(stashDir, ".akm", "runs");
-    if (fs.existsSync(runsDir)) {
-      const entries = fs.readdirSync(runsDir);
-      expect(entries.length).toBe(0);
-    }
-
-    // Stderr should NOT contain the "improve result written to" hint.
-    expect(result.stderr).not.toContain("improve result written to");
-  });
-
-  test("two consecutive default-mode runs produce distinct improve_runs rows", () => {
-    const a = runCli(["improve", "--dry-run"], stashDir);
-    expect(a.status).toBe(0);
-    expect(a.stdout).toBe("");
-
-    // Use the same xdgData root so the second run accumulates into the
-    // same state.db. runCli() mints a fresh xdgData per call, so we re-run
-    // by passing the same env via a small inline driver.
-    const b = runCli(["improve", "--dry-run"], stashDir);
-    expect(b.status).toBe(0);
-    expect(b.stdout).toBe("");
-
-    // Each run wrote a row into its own state.db (one row per state.db).
-    const aRows = readImproveRuns(a.xdgData);
-    const bRows = readImproveRuns(b.xdgData);
-    expect(aRows.length).toBe(1);
-    expect(bRows.length).toBe(1);
-    expect(aRows[0].id).not.toEqual(bRows[0].id);
-    // No legacy directory under either stash root.
-    expect(fs.existsSync(path.join(stashDir, ".akm", "runs"))).toBe(false);
   });
 });
