+++ tests/env-path-run.test.ts
@@ -1,15 +1,15 @@
 /**
- * `akm env path` / `env export` / `env run` CLI behavior.
+ * `akm env path` / `env export` / `env run` CLI behavior — in-process only.
  *
- * `env run` MUST keep spawning a real subprocess: the CLI spawns the target
- * command with stdout/stderr inherited to the real file descriptors, so the
- * injected-env output is the child's, not the parent's, and the in-process
- * console-capture harness cannot observe it. A real process boundary is the
- * whole point. Pure path/export resolution runs in-process.
+ * All tests here run through the in-process `runCliCapture` harness: pure
+ * path/export resolution plus `env run` error paths that fail BEFORE any
+ * child process is spawned. The `env run` / `secret run` happy paths that
+ * actually spawn a target command (whose fd-inherited stdout is the
+ * contract) live in tests/integration/env-run.test.ts — only a real process
+ * boundary can observe the child's output.
  */
 
 import { afterAll, afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { spawnSync } from "node:child_process";
 import fs from "node:fs";
 import path from "node:path";
 import { resetGraphBoostCache } from "../src/indexer/graph/graph-boost";
@@ -23,9 +23,6 @@ afterAll(() => {
   for (const d of disposers) d.cleanup();
   disposers.length = 0;
 });
-
-const repoRoot = path.resolve(import.meta.dir, "..");
-const cliPath = path.join(repoRoot, "src", "cli.ts");
 
 function makeStash(): string {
   const stash = makeStashDir();
@@ -44,29 +41,6 @@ async function runCli(
     const { stdout, stderr, code } = await runCliCapture(args);
     return { stdout, stderr, status: code };
   });
-}
-
-function spawnCli(
-  args: string[],
-  extraEnv: Record<string, string | undefined> = {},
-  stdinInput?: string,
-): { stdout: string; stderr: string; status: number } {
-  const result = spawnSync("bun", [cliPath, ...args], {
-    encoding: "utf8",
-    timeout: 15_000,
-    cwd: repoRoot,
-    input: stdinInput,
-    env: {
-      ...process.env,
-      AKM_STASH_DIR: undefined,
-      ...extraEnv,
-    },
-  });
-  return {
-    stdout: result.stdout ?? "",
-    stderr: result.stderr ?? "",
-    status: result.status ?? 1,
-  };
 }
 
 beforeEach(() => {
@@ -133,56 +107,8 @@ describe("env export", () => {
 });
 
 describe("env run", () => {
-  // KEPT AS A SUBPROCESS: `env run` spawns the target with stdout inherited to
-  // the real fd; the injected-env output is the child's, not visible in-process.
-  test("runs a command with the whole env file injected", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "FOO=bar\nBAR=baz\n", "utf8");
-
-    const { stdout, stderr, status } = spawnCli(
-      ["env", "run", "env:prod", "--", "bash", "-lc", 'printf \'%s %s\' "$FOO" "$BAR"'],
-      { AKM_STASH_DIR: stashDir },
-    );
-
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("bar baz");
-    expect(stderr.trim()).toBe("");
-  });
-
-  test("substitutes ${secret:NAME} tokens with the sibling secret value", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.mkdirSync(path.join(stashDir, "secrets"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "secrets", "my_api_token"), "s3cr3t", "utf8");
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "API_KEY=Bearer ${secret:my_api_token}\n", "utf8");
-
-    const { stdout, status } = spawnCli(["env", "run", "env:prod", "--", "bash", "-lc", "printf '%s' \"$API_KEY\""], {
-      AKM_STASH_DIR: stashDir,
-    });
-
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("Bearer s3cr3t");
-  });
-
-  test("substitutes multiple tokens embedded in a value and leaves ${HOME} untouched", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.mkdirSync(path.join(stashDir, "secrets"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "secrets", "a"), "AAA", "utf8");
-    fs.writeFileSync(path.join(stashDir, "secrets", "b"), "BBB", "utf8");
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "PAIR=${secret:a}:${secret:b}\nKEEP=${HOME}\n", "utf8");
-
-    const { stdout, status } = spawnCli(
-      ["env", "run", "env:prod", "--", "bash", "-lc", 'printf \'%s|%s\' "$PAIR" "$KEEP"'],
-      { AKM_STASH_DIR: stashDir },
-    );
-
-    expect(status).toBe(0);
-    // PAIR fully substituted; KEEP left as the literal token (no secret named HOME).
-    expect(stdout.trim()).toBe("AAA:BBB|${HOME}");
-  });
-
+  // Only pre-spawn error paths run here; the spawn-and-observe-child-output
+  // tests live in tests/integration/env-run.test.ts.
   test("exits non-zero and injects nothing when a referenced secret is missing", async () => {
     const stashDir = makeStash();
     fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
@@ -216,92 +142,6 @@ describe("env run", () => {
     expect(parsed.error).toContain("akm secret run");
   });
 
-  test("warns but injects when a first-party env file contains a hijack var", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    // EDITOR is on the dangerous-key list (RCE vector when sourced from an
-    // untrusted stash) but, unlike PATH, does not break command resolution —
-    // so this exercises the warn-and-inject path for a first-party stash.
-    fs.writeFileSync(path.join(stashDir, "env", "danger.env"), "EDITOR=/evil\nFOO=ok\n", "utf8");
-
-    const { stdout, stderr, status } = spawnCli(
-      ["env", "run", "env:danger", "--", "bash", "-lc", "printf '%s' \"$FOO\""],
-      { AKM_STASH_DIR: stashDir },
-    );
-
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("ok");
-    expect(stderr).toContain("EDITOR");
-  });
-
-  test("--only injects just the named keys", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "FOO=foo\nBAR=bar\nBAZ=baz\n", "utf8");
-
-    const { stdout, status } = spawnCli(
-      ["env", "run", "env:prod", "--only", "FOO,BAZ", "--", "bash", "-lc", 'printf \'%s|%s|%s\' "$FOO" "$BAR" "$BAZ"'],
-      { AKM_STASH_DIR: stashDir },
-    );
-
-    expect(status).toBe(0);
-    // FOO and BAZ injected; BAR excluded.
-    expect(stdout.trim()).toBe("foo||baz");
-  });
-
-  test("--except injects all but the named keys", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "FOO=foo\nBAR=bar\n", "utf8");
-
-    const { stdout, status } = spawnCli(
-      ["env", "run", "env:prod", "--except", "BAR", "--", "bash", "-lc", 'printf \'%s|%s\' "$FOO" "$BAR"'],
-      { AKM_STASH_DIR: stashDir },
-    );
-
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("foo|");
-  });
-
-  test("--clean does not inherit unrelated parent env vars", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "FOO=foo\n", "utf8");
-
-    const { stdout, status } = spawnCli(
-      ["env", "run", "env:prod", "--clean", "--", "bash", "-lc", 'printf "%s|%s" "$FOO" "${PARENT_ONLY:-}"'],
-      { AKM_STASH_DIR: stashDir, PARENT_ONLY: "sentinel" },
-    );
-
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("foo|");
-  });
-
-  test("--clean --inherit passes through selected parent env vars", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "FOO=foo\n", "utf8");
-
-    const { stdout, status } = spawnCli(
-      [
-        "env",
-        "run",
-        "env:prod",
-        "--clean",
-        "--inherit",
-        "PARENT_ONLY",
-        "--",
-        "bash",
-        "-lc",
-        'printf "%s|%s" "$FOO" "${PARENT_ONLY:-}"',
-      ],
-      { AKM_STASH_DIR: stashDir, PARENT_ONLY: "sentinel" },
-    );
-
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("foo|sentinel");
-  });
-
   test("--only and --except together is rejected", async () => {
     const stashDir = makeStash();
     fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
@@ -320,42 +160,19 @@ describe("env run", () => {
 });
 
 describe("vault run (removed in 0.9.0)", () => {
-  // KEPT AS A SUBPROCESS: child-process stdout boundary, same as env run.
-  test("the `akm vault` verb no longer exists", () => {
+  test("the `akm vault` verb no longer exists", async () => {
     const stashDir = makeStash();
     fs.mkdirSync(path.join(stashDir, "env"), { recursive: true });
     fs.writeFileSync(path.join(stashDir, "env", "prod.env"), "FOO=bar\nBAR=baz\n", "utf8");
 
-    const { status } = spawnCli(["vault", "run", "vault:prod", "--", "bash", "-lc", 'printf \'%s %s\' "$FOO" "$BAR"'], {
-      AKM_STASH_DIR: stashDir,
-    });
-
-    expect(status).not.toBe(0);
-  });
-});
-
-describe("secret run", () => {
-  test("--clean does not inherit unrelated parent env vars", () => {
-    const stashDir = makeStash();
-    fs.mkdirSync(path.join(stashDir, "secrets"), { recursive: true });
-    fs.writeFileSync(path.join(stashDir, "secrets", "token"), "sekret", "utf8");
-
-    const { stdout, status } = spawnCli(
-      [
-        "secret",
-        "run",
-        "secret:token",
-        "API_TOKEN",
-        "--clean",
-        "--",
-        "bash",
-        "-lc",
-        'printf "%s|%s" "$API_TOKEN" "${PARENT_ONLY:-}"',
-      ],
-      { AKM_STASH_DIR: stashDir, PARENT_ONLY: "sentinel" },
+    const { status } = await runCli(
+      ["vault", "run", "vault:prod", "--", "bash", "-lc", 'printf \'%s %s\' "$FOO" "$BAR"'],
+      {
+        AKM_STASH_DIR: stashDir,
+      },
     );
 
-    expect(status).toBe(0);
-    expect(stdout.trim()).toBe("sekret|");
+    // citty exits non-zero for an unknown top-level command.
+    expect(status).not.toBe(0);
   });
 });
