+++ tests/remember-frontmatter.test.ts
@@ -12,7 +12,6 @@
  */
 
 import { afterEach, describe, expect, test } from "bun:test";
-import { spawnSync } from "node:child_process";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
@@ -25,20 +24,17 @@ import { withEnv } from "./_helpers/sandbox";
 
 // ── CLI harness ──────────────────────────────────────────────────────────────
 //
-// Migrated the non-stdin `akm remember` invocations from spawnSync to the shared
-// in-process harness (tests/_helpers/cli.ts). `remember` resolves its target
-// from AKM_STASH_DIR (XDG), not process.cwd(), so it runs faithfully in-process.
+// All `akm remember` invocations here run through the shared in-process harness
+// (tests/_helpers/cli.ts). `remember` resolves its target from AKM_STASH_DIR
+// (XDG), not process.cwd(), so it runs faithfully in-process.
 //
-// KEPT SPAWNING (harness gap): the two tests that pipe a body via stdin
-// (`spawnRunCli({ input })`). runCliCapture has no stdin support — the CLI's
-// `remember` reads process.stdin when no body arg is given (and when --format
-// json is present), which the in-process harness cannot supply. Those two tests
-// stay as real subprocesses so the stdin read path is exercised faithfully.
+// The two stdin-driven tests (piping a body via a real subprocess's stdin) live
+// in tests/integration/remember-stdin.test.ts — runCliCapture has no stdin
+// support and the CLI's process.stdin read path is the thing under test there.
 //
 // Env mutation goes through the allowlisted withEnv wrapper; temp dirs are
 // created via makeTempDir (kept local) and tracked in tempDirs for cleanup.
 
-const CLI = path.join(__dirname, "..", "src", "cli.ts");
 const tempDirs: string[] = [];
 
 function makeTempDir(prefix: string): string {
@@ -68,21 +64,6 @@ async function runCli(args: string[], options?: { stashDir?: string }) {
   return { stashDir, result: { status: code, stdout, stderr } };
 }
 
-/**
- * Subprocess runner, retained ONLY for the stdin-driven tests. runCliCapture has
- * no stdin support, so the CLI's stdin read path is exercised via a real process.
- */
-function spawnRunCli(args: string[], options?: { stashDir?: string; input?: string }) {
-  const { stashDir, env } = freshDirs(options);
-  const result = spawnSync("bun", [CLI, ...args], {
-    encoding: "utf8",
-    timeout: 30_000,
-    input: options?.input,
-    env: { ...process.env, ...env },
-  });
-  return { stashDir, result };
-}
-
 afterEach(() => {
   for (const dir of tempDirs.splice(0)) {
     fs.rmSync(dir, { recursive: true, force: true });
@@ -110,25 +91,8 @@ describe("zero-flag remember", () => {
     expect(stashDir).toBeTruthy();
   });
 
-  test("stdin zero-flag path also writes captureMode: hot + beliefState: asserted", () => {
-    const { result } = spawnRunCli(["remember"], { input: "VPN needed for staging deploys" });
-    expect(result.status).toBe(0);
-    const json = JSON.parse(result.stdout) as { ref: string; path: string };
-    const content = fs.readFileSync(json.path, "utf8");
-    expect(content.startsWith("---")).toBe(true);
-    const parsed = parseFrontmatter(content);
-    expect(parsed.data.captureMode).toBe("hot");
-    expect(parsed.data.beliefState).toBe("asserted");
-    expect(Object.keys(parsed.data).sort()).toEqual(["beliefState", "captureMode"]);
-  });
-
-  test("reads stdin when --format json is present", () => {
-    const { result } = spawnRunCli(["remember", "--name", "from-stdin", "--format", "json"], { input: "stdin body" });
-    expect(result.status).toBe(0);
-    const json = JSON.parse(result.stdout) as { path: string };
-    expect(fs.readFileSync(json.path, "utf8")).toContain("stdin body");
-    expect(fs.readFileSync(json.path, "utf8")).not.toContain("\njson");
-  });
+  // The stdin variants of these tests (piping a body via a real subprocess)
+  // live in tests/integration/remember-stdin.test.ts.
 });
 
 // ── CLI args (Mode 1) ────────────────────────────────────────────────────────
